requirements.txt
gradio[oauth]==4.44.1
requests
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
langchain-tavily
langchain-chroma
langgraph
huggingface_hub
sentence-transformers
arxiv
pymupdf
wikipedia
pgvector
python-dotenv
protobuf==3.20.*
chromadb
langchain-openai
openai